files/programs/board_config/program/requirements.txt
PyYaml
paho-mqtt
